begins migrating tabs
