+++ packages/ui/src/forms/field-types/Tabs/index.tsx
@@ -1,72 +1,17 @@
-'use client'
-import React, { useCallback, useEffect, useState } from 'react'
-import { useTranslation } from '../../../providers/Translation'
+import React from 'react'
 
-import type { Tab } from 'payload/types'
-import type { DocumentPreferences } from 'payload/types'
 import type { Props } from './types'
 
 import { tabHasName } from 'payload/types'
-import { getTranslation } from '@payloadcms/translations'
-import { toKebabCase } from 'payload/utilities'
-import { useCollapsible } from '../../../elements/Collapsible/provider'
-import { ErrorPill } from '../../../elements/ErrorPill'
-import { useDocumentInfo } from '../../../providers/DocumentInfo'
-import { usePreferences } from '../../../providers/Preferences'
 import FieldDescription from '../../FieldDescription'
-import { useFormSubmitted } from '../../Form/context'
 import { createNestedFieldPath } from '../../Form/createNestedFieldPath'
 import RenderFields from '../../RenderFields'
-import { WatchChildErrors } from '../../WatchChildErrors'
-import { fieldBaseClass } from '../shared'
 import './index.scss'
 import { TabsProvider } from './provider'
+import { TabComponent } from './Tab'
+import { Wrapper } from './Wrapper'
 
 const baseClass = 'tabs-field'
-
-type TabProps = {
-  isActive?: boolean
-  parentPath: string
-  setIsActive: () => void
-  tab: Tab
-}
-
-const TabComponent: React.FC<TabProps> = ({ isActive, parentPath, setIsActive, tab }) => {
-  const { i18n } = useTranslation()
-  const [errorCount, setErrorCount] = useState(undefined)
-  const hasName = tabHasName(tab)
-  const submitted = useFormSubmitted()
-
-  const pathSegments = []
-  if (parentPath) pathSegments.push(parentPath)
-  if (hasName) pathSegments.push(tab.name)
-  const path = pathSegments.join('.')
-  const tabHasErrors = submitted && errorCount > 0
-
-  return (
-    <React.Fragment>
-      <WatchChildErrors
-        fieldSchema={hasName ? undefined : tab.fields}
-        path={path}
-        setErrorCount={setErrorCount}
-      />
-      <button
-        className={[
-          `${baseClass}__tab-button`,
-          tabHasErrors && `${baseClass}__tab-button--has-error`,
-          isActive && `${baseClass}__tab-button--active`,
-        ]
-          .filter(Boolean)
-          .join(' ')}
-        onClick={setIsActive}
-        type="button"
-      >
-        {tab.label ? getTranslation(tab.label, i18n) : hasName && tab.name}
-        {tabHasErrors && <ErrorPill count={errorCount} />}
-      </button>
-    </React.Fragment>
-  )
-}
 
 const TabsField: React.FC<Props> = (props) => {
   const {
@@ -77,72 +22,62 @@ const TabsField: React.FC<Props> = (props) => {
     path,
     permissions,
     tabs,
+    formState,
+    user,
   } = props
 
-  const { getPreference, setPreference } = usePreferences()
-  const { preferencesKey } = useDocumentInfo()
-  const { i18n } = useTranslation()
-
-  const isWithinCollapsible = useCollapsible()
-  const [activeTabIndex, setActiveTabIndex] = useState<number>(0)
   const tabsPrefKey = `tabs-${indexPath}`
 
-  useEffect(() => {
-    const getInitialPref = async () => {
-      const existingPreferences: DocumentPreferences = await getPreference(preferencesKey)
-      const initialIndex = path
-        ? existingPreferences?.fields?.[path]?.tabIndex
-        : existingPreferences?.fields?.[tabsPrefKey]?.tabIndex
-      setActiveTabIndex(initialIndex || 0)
-    }
-    void getInitialPref()
-  }, [path, indexPath, getPreference, preferencesKey, tabsPrefKey])
+  // useEffect(() => {
+  //   const getInitialPref = async () => {
+  //     const existingPreferences: DocumentPreferences = await getPreference(preferencesKey)
+  //     const initialIndex = path
+  //       ? existingPreferences?.fields?.[path]?.tabIndex
+  //       : existingPreferences?.fields?.[tabsPrefKey]?.tabIndex
+  //     setActiveTabIndex(initialIndex || 0)
+  //   }
+  //   void getInitialPref()
+  // }, [path, indexPath, getPreference, preferencesKey, tabsPrefKey])
 
-  const handleTabChange = useCallback(
-    async (incomingTabIndex: number) => {
-      setActiveTabIndex(incomingTabIndex)
+  // const handleTabChange = useCallback(
+  //   async (incomingTabIndex: number) => {
+  //     setActiveTabIndex(incomingTabIndex)
 
-      const existingPreferences: DocumentPreferences = await getPreference(preferencesKey)
+  //     const existingPreferences: DocumentPreferences = await getPreference(preferencesKey)
 
-      setPreference(preferencesKey, {
-        ...existingPreferences,
-        ...(path
-          ? {
-              fields: {
-                ...(existingPreferences?.fields || {}),
-                [path]: {
-                  ...existingPreferences?.fields?.[path],
-                  tabIndex: incomingTabIndex,
-                },
-              },
-            }
-          : {
-              fields: {
-                ...existingPreferences?.fields,
-                [tabsPrefKey]: {
-                  ...existingPreferences?.fields?.[tabsPrefKey],
-                  tabIndex: incomingTabIndex,
-                },
-              },
-            }),
-      })
-    },
-    [preferencesKey, getPreference, setPreference, path, tabsPrefKey],
-  )
+  //     setPreference(preferencesKey, {
+  //       ...existingPreferences,
+  //       ...(path
+  //         ? {
+  //             fields: {
+  //               ...(existingPreferences?.fields || {}),
+  //               [path]: {
+  //                 ...existingPreferences?.fields?.[path],
+  //                 tabIndex: incomingTabIndex,
+  //               },
+  //             },
+  //           }
+  //         : {
+  //             fields: {
+  //               ...existingPreferences?.fields,
+  //               [tabsPrefKey]: {
+  //                 ...existingPreferences?.fields?.[tabsPrefKey],
+  //                 tabIndex: incomingTabIndex,
+  //               },
+  //             },
+  //           }),
+  //     })
+  //   },
+  //   [preferencesKey, getPreference, setPreference, path, tabsPrefKey],
+  // )
+
+  // TODO: wire this in
+  const activeTabIndex = 0
 
   const activeTabConfig = tabs[activeTabIndex]
 
   return (
-    <div
-      className={[
-        fieldBaseClass,
-        className,
-        baseClass,
-        isWithinCollapsible && `${baseClass}--within-collapsible`,
-      ]
-        .filter(Boolean)
-        .join(' ')}
-    >
+    <Wrapper className={className}>
       <TabsProvider>
         <div className={`${baseClass}__tabs-wrap`}>
           <div className={`${baseClass}__tabs`}>
@@ -152,7 +87,8 @@ const TabsField: React.FC<Props> = (props) => {
                   isActive={activeTabIndex === tabIndex}
                   key={tabIndex}
                   parentPath={path}
-                  setIsActive={() => handleTabChange(tabIndex)}
+                  setIsActive={undefined}
+                  // setIsActive={() => handleTabChange(tabIndex)}
                   tab={tab}
                 />
               )
@@ -165,10 +101,10 @@ const TabsField: React.FC<Props> = (props) => {
               <div
                 className={[
                   `${baseClass}__tab`,
-                  activeTabConfig.label &&
-                    `${baseClass}__tabConfigLabel-${toKebabCase(
-                      getTranslation(activeTabConfig.label, i18n),
-                    )}`,
+                  // activeTabConfig.label &&
+                  //   `${baseClass}__tabConfigLabel-${toKebabCase(
+                  //     getTranslation(activeTabConfig.label, i18n),
+                  //   )}`,
                 ]
                   .filter(Boolean)
                   .join(' ')}
@@ -179,8 +115,7 @@ const TabsField: React.FC<Props> = (props) => {
                   marginPlacement="bottom"
                   path={path}
                 />
-                [RenderFields]
-                {/* <RenderFields
+                <RenderFields
                   fieldSchema={activeTabConfig.fields.map((field) => {
                     const pathSegments = []
 
@@ -197,7 +132,7 @@ const TabsField: React.FC<Props> = (props) => {
                   indexPath={indexPath}
                   key={
                     activeTabConfig.label
-                      ? getTranslation(activeTabConfig.label, i18n)
+                      ? activeTabConfig.label // ? getTranslation(activeTabConfig.label, i18n)
                       : activeTabConfig['name']
                   }
                   margins="small"
@@ -207,13 +142,15 @@ const TabsField: React.FC<Props> = (props) => {
                       : permissions
                   }
                   readOnly={readOnly}
-                /> */}
+                  user={user}
+                  formState={formState}
+                />
               </div>
             </React.Fragment>
           )}
         </div>
       </TabsProvider>
-    </div>
+    </Wrapper>
   )
 }
