+++ packages/ui/src/forms/field-types/Tabs/provider.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { createContext, useContext } from 'react'
 
 const Context = createContext(false)
